requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
bcrypt==4.1.2
pymongo==4.6.1
python-dotenv==1.0.1